- Теперь получаем информачию через get метод - Убрал хард код и заменил на jsp страницы - Заменил "/WEB-INF/changeLogin.jsp" на "/WEB-INF/getUser.jsp". (случайно не поменял, когда брал код с другого сервлета.)
