+++ src/main/java/com/teachmeskills/hw_35/database/PostgresDriverManager.java
@@ -1,4 +1,4 @@
-package com.teachmeskills.hw_35.database;
+package com.teachmeskills.database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
