+++ src/main/java/com/teachmeskills/pack/InnerClass.java
@@ -1,0 +1,13 @@
+//package com.teachmeskills.pack;
+//
+//public class InnerClass {
+//    private String innerMessage;
+//
+//    public void setInnerMessage() {
+//        System.out.println(innerMessage);
+//    }
+//
+//    public void setInnerMessage(String innerMessage) {
+//        this.innerMessage = innerMessage;
+//    }
+//}
