+++ src/main/java/com/teachmeskills/pack/SpringHelloWorld.java
@@ -1,0 +1,24 @@
+//package com.teachmeskills.pack;
+//
+//
+//public class SpringHelloWorld {
+//
+//    private String message;
+//    private InnerClass innerClass;
+//
+//    public void getMessage() {
+//        System.out.println(message);
+//    }
+//
+//    public void setMessage(String message) {
+//        this.message = message;
+//    }
+//
+//    public InnerClass getInnerClass() {
+//        return innerClass;
+//    }
+//
+//    public void setInnerClass(InnerClass innerClass){
+//        this.innerClass = innerClass;
+//    }
+//}
