+++ src/main/resources/my-config.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="
+        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+
+    <bean id="inner" class="com.teachmeskills.pack.InnerClass" scope="prototype">
+        <property name="innerMessage" value="Inner!!!"/>
+    </bean>
+
+    <bean id="aaa" class="com.teachmeskills.pack.SpringHelloWorld" scope="prototype">
+        <property name="innerClass" ref="inner"/>
+        <property name="message" value="TEST"/>
+    </bean>
+
+</beans>
