+++ src/main/webapp/WEB-INF/error.jsp
@@ -1,0 +1,18 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: PC
+  Date: 13.08.2024
+  Time: 15:23
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title>Error</title>
+</head>
+<body>
+<h1>Error insert</h1>
+<a href="main">Go to the main page</a>
+</br>
+</body>
+</html>
