+++ src/main/webapp/WEB-INF/info.jsp
@@ -1,0 +1,18 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: PC
+  Date: 13.08.2024
+  Time: 15:26
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title>info</title>
+</head>
+<body>
+<p>Login: ${login}</p>
+<p>Name: ${name}</p>
+<a href="main">Go to the main page</a>
+</body>
+</html>
