+++ src/main/webapp/WEB-INF/successInsert.jsp
@@ -1,0 +1,18 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: PC
+  Date: 13.08.2024
+  Time: 17:42
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title>Success insert</title>
+</head>
+<body>
+<h1>Success insert</h1>
+<a href="main">Go to the main page</a>
+</br>
+</body>
+</html>
